+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/mode/simple/conversion/model/SimpleConversionEvent.kt
@@ -1,0 +1,10 @@
+package io.github.composegears.valkyrie.ui.screen.mode.simple.conversion.model
+
+sealed interface SimpleConversionEvent {
+    data class ExportKtFile(
+        val fileName: String,
+        val content: String,
+    ) : SimpleConversionEvent
+
+    data class CopyInClipboard(val text: String) : SimpleConversionEvent
+}