+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/tools/imagevectorxml/conversion/model/ImageVectorXmlEvent.kt
@@ -1,0 +1,10 @@
+package io.github.composegears.valkyrie.ui.screen.tools.imagevectorxml.conversion.model
+
+sealed interface ImageVectorXmlEvent {
+    data class ExportXmlFile(
+        val fileName: String,
+        val content: String,
+    ) : ImageVectorXmlEvent
+
+    data class CopyInClipboard(val text: String) : ImageVectorXmlEvent
+}